Activate user (not finished)
Started to implement admin feature to activate/deactivate users. My plan was to have a button on admin page for this option. When pressed, opens a window with text box to type user in and this will search users and then make self.is_active == False or True. Lemme know on instagram if there is a better way to approach this. Will work on it tomorrow